+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Changes from previous releases are listed below.
 - Renaming and moving of Sentinel-2 bands _(see #42)_
 - Process clipping of Sentinel tiles on monthly basis _(see #29)_
 - Adjust the shapefile CRS transformation for Sentinel-1 _(see #49)_
+- Adjusting split generation _(see #45)_
+- Fixing padding mask _(see #50)_
 
 ## 0.3.1 (2024-07-29)
 - Remove country_code variable in collector downloader _(see #33)_